Create Site API Module.
A new form text field named 'Site API Key'  added to the 'Site Information' form with the default value of 'No API Key yet'.
When this form is submitted, the value that the user entered for this field should be saved as the system variable named 'siteapikey'.
The text of the 'Save configuration' button change to 'Update Configuration'.